changed the three int to double

+++ Chapter3/threedashnineteen.java
@@ -15,17 +15,17 @@ public class threedashnineteen {
 			//prompt the user to enter side1
 			System.out.print("Enter length of side 1: ") ;
 			//reads an input from the user in a double value from the console as side1
-			int side1 = input.nextInt() ;
+			double side1 = input.nextInt() ;
 			
 			//prompt the user to enter side2
 			System.out.print("Enter length of side 2: ") ;
 			//reads an input from the user in a double value from the console as side2
-			int side2 = input.nextInt() ;
+			double side2 = input.nextInt() ;
 			
 			//prompt the user to enter side3
 			System.out.print("Enter length of side 3: ") ;
 			//reads an input from the user in a double value from the console as side3
-			int side3 = input.nextInt() ;
+			double side3 = input.nextInt() ;
 			
 			//check validity of inputs from user
 			if (side3 <= side1 + side2) { ;
